ui components from SettingsScreen.kt moved to ui.settings package
