+++ app/src/main/java/srimani7/apps/feedfly/ui/settings/SwitchPreference.kt
@@ -1,0 +1,21 @@
+package srimani7.apps.feedfly.ui.settings
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Switch
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun SwitchPreference(text: String, checked: Boolean, onChange: (Boolean) -> Unit) {
+    TextButton(onClick = { onChange(!checked) }) {
+        Text(text = text, style = MaterialTheme.typography.bodyLarge)
+        Spacer(modifier = Modifier.weight(1f))
+        Switch(
+            checked = checked,
+            onCheckedChange = onChange
+        )
+    }
+}
